document: major refactor
We got a problem with document, the whole 'changeFieldValue' was very
clumsy and because of this, the code was complex and hard to test.

They way documents are tracked for change is now based on a field
called revision. It's calculated one time when the request comes and
then it's propagated for everywhere.

Documents don't get deleted and the trigger messages instead of
carrying the whole document, now, use only it's id and revision.

Resolves: #50, #66, #70, #74, #75