domain/document: add fallback during resource fetch
The second time a document with the same id is received, the old
version contains the id of the resource. But, the resource, could be
deleted and recreated with another id. So, if we don't find the
resource by it's id, we fallback to the find using the document uri.